Fix dangling pointer issue in buffer
When a linegroup was cleared, it was not properly removed from the list.
This was hidden when the linegroup was at the end of the buffer, as
anchor pointers were updated, but when lines in the middle of the buffer
were removed, this resulted in list entries being freed without being
unlinked first.